Expire otp and password attempts

+++ budgetapp-server/src/env.rs
@@ -47,9 +47,9 @@ pub struct Lifetimes {
 #[derive(Deserialize, Serialize)]
 pub struct Security {
     pub otp_max_attempts: i16,
-    pub otp_attempts_reset_mins: i16,
+    pub otp_attempts_reset_mins: u64,
     pub password_max_attempts: i16,
-    pub password_attempts_reset_mins: i16,
+    pub password_attempts_reset_mins: u64,
     pub password_min_len_chars: usize,
 }
